fix supervision for usb dump
