+++ lib/platform/app/sync/usb_drive_dump/completer.ex
@@ -1,0 +1,16 @@
+defmodule Platform.App.Sync.UsbDriveDump.Completer do
+  @moduledoc "Finishes files dump into room"
+
+  use GracefulGenServer
+
+  alias Chat.Sync.UsbDriveDumpRoom
+
+  @impl true
+  def on_init(_args) do
+    UsbDriveDumpRoom.complete()
+  end
+
+  @impl true
+  def on_exit(_reason, _state) do
+  end
+end
