+++ lib/platform/app/sync/usb_drive_dump_supervisor.ex
@@ -1,12 +1,13 @@
 defmodule Platform.App.Sync.UsbDriveDumpSupervisor do
   use Supervisor
+  import Platform
 
   require Logger
 
-  alias Platform.App.Sync.UsbDriveDump.Logic
+  alias Platform.App.Sync.UsbDriveDump.Completer
+  alias Platform.App.Sync.UsbDriveDump.Dumper
   alias Platform.App.Sync.UsbDriveDumpSupervisor.Tasks
   alias Platform.Storage.Backup.Starter
-  alias Platform.Storage.MountedHealer
 
   @mount_path Application.compile_env(:platform, :mount_path_media)
 
@@ -15,17 +16,26 @@ defmodule Platform.App.Sync.UsbDriveDumpSupervisor do
   end
 
   @impl Supervisor
-  def init([device]) do
+  def init([_device]) do
     "UsbDriveDumpSupervisor start" |> Logger.info()
 
     full_path = [@mount_path, "DCIM"] |> Path.join()
     tasks = Tasks
+    files_dumped_stage = Platform.App.Sync.UsbDriveDump.FilesDumpedStage
 
     [
-      {Task.Supervisor, name: tasks},
-      # {MountedHealer, [device, full_path, tasks]},
+      use_task(tasks),
       {Starter, flag: :usb_drive_dump},
-      {Logic, [full_path, tasks]}
+      use_next_stage(files_dumped_stage),
+      {Dumper,
+       mounted: full_path,
+       task_in: tasks,
+       next: [
+         under: files_dumped_stage,
+         run: [
+           Completer
+         ]
+       ]}
     ]
     |> Supervisor.init(strategy: :rest_for_one)
     |> tap(fn res ->
